:recycle: Split DBManager
